+++ src/Internal/PdfPageInfo.php
@@ -30,8 +30,7 @@ final class PdfPageInfo extends AbstractPdfNumber
     public function __construct(
         public ?PdfRotation $rotation = null,
         public ?PdfSize $size = null,
-    ) {
-    }
+    ) {}
 
     /**
      * @phpstan-assert-if-true PdfRotation $this->rotation