+++ src/PdfDocument.php
@@ -690,9 +690,7 @@ class PdfDocument
      *
      * @see PdfDocument::header()
      */
-    public function footer(): void
-    {
-    }
+    public function footer(): void {}
 
     /**
      * Gets the alias for the total number of pages.
@@ -940,7 +938,7 @@ class PdfDocument
         /** @var int[] $keys */
         $keys = \unpack('C*', $str);
         $charWidths = $this->currentFont->cw;
-        $width = \array_sum(\array_map(static fn (int $key): int => $charWidths[$key], $keys));
+        $width = \array_sum(\array_map(static fn(int $key): int => $charWidths[$key], $keys));
 
         return (float) $width * $this->fontSize / 1000.0;
     }
@@ -1004,9 +1002,7 @@ class PdfDocument
      *
      * @see PdfDocument::footer()
      */
-    public function header(): void
-    {
-    }
+    public function header(): void {}
 
     /**
      * Draws a horizontal line with current draw color and optionally the given line width.
@@ -2975,7 +2971,7 @@ class PdfDocument
         if ($image->isTransparencies()) {
             $transparencies = \array_reduce(
                 $image->transparencies,
-                static fn (string $carry, int $value): string => $carry . \sprintf('%1$d %1$d ', $value),
+                static fn(string $carry, int $value): string => $carry . \sprintf('%1$d %1$d ', $value),
                 ''
             );
             $this->putf('/Mask [%s]', $transparencies);